+++ app/types/walletExportTypes.js
@@ -1,0 +1,2 @@
+// @flow
+export type walletExportTypeChoices = 'full' | 'readOnly' | 'paperWallet';